refactor(arch): extract rank_fusion sub-module from ranked_context.rs (PR-H)
ranked_context.rs had grown to 958 LOC because stage-4 rank-fusion
(5 helpers + RankFusionPolicy struct + rank_fusion_policy fn) lived
alongside the stage orchestration in run_ranked_context. PR-H pulls
the stage-4 region into a sibling sub-module.

New `symbol_query/rank_fusion.rs` (323 LOC) owns:
  - RankFusionPolicy + rank_fusion_policy (query-shape-tuned thresholds)
  - merge_semantic_ranked_entries
  - merge_sparse_ranked_entries
  - compact_semantic_evidence
  - compact_sparse_evidence
  - annotate_ranked_context_provenance

Visibility: every export is `pub(super)`. ranked_context.rs is the
only legitimate caller — these helpers don't make sense in isolation
from the pipeline's stage ordering. Tests live where they were
(ranked_context.rs `mod tests`) because they exercise the
helpers through the same scope they'd be invoked from in production.

ranked_context.rs: 958 → 665 LOC (-31%).

Validation:
- cargo clippy --workspace --features http,semantic -- -D warnings: clean
- cargo clippy --workspace --no-default-features -- -D warnings: clean
- cargo clippy --workspace --features scip-backend -- -D warnings: clean
- cargo test -p codelens-mcp --bin codelens-mcp: 593 passed / 0 failed
- cargo fmt --all -- --check: clean
